opt: abstract web search interface, code cleanup
